Hard-coded addressses
All radiance functions are not currently bieng found by my machine, so hard-coded the location of the commands. I assume this can be fixed by knowing how to modify my path... but at the moment I'm trying to get the work to code.

Still stuck on _pOpen altough I got oconv to work once it seems...